+++ StraviaTECCore/StraviaTECCore/Models/Gestionactividad.cs
@@ -5,7 +5,7 @@ namespace StraviaTECCore.Models
 {
     public partial class Gestionactividad
     {
-        public decimal GestionId { get; set; }
+        public int GestionId { get; set; }
         public string Tipo { get; set; }
         public string Clasificacion { get; set; }
